+++ libs/components/validation/projects/validation/documentation/code-examples/url-validation/directive/url-validation-demo.module.ts
@@ -1,0 +1,44 @@
+import {
+  CommonModule
+} from '@angular/common';
+
+import {
+  NgModule
+} from '@angular/core';
+
+import {
+  FormsModule
+} from '@angular/forms';
+
+import {
+  SkyIdModule
+} from '@skyux/core';
+
+import {
+  SkyInputBoxModule
+} from '@skyux/forms';
+
+import {
+  SkyUrlValidationModule
+} from 'projects/validation/src/public-api';
+
+import {
+  UrlValidationDemoComponent
+} from './url-validation-demo.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    FormsModule,
+    SkyIdModule,
+    SkyInputBoxModule,
+    SkyUrlValidationModule
+  ],
+  exports: [
+    UrlValidationDemoComponent
+  ],
+  declarations: [
+    UrlValidationDemoComponent
+  ]
+})
+export class UrlValidationDemoModule { }