+++ libs/components/validation/projects/validation/src/modules/validators/validators.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { SkyValidators } from './validators';
+
+
+
+@NgModule({
+  exports: [SkyValidators]
+})
+export class ValidatorsModule { }