+++ libs/components/validation/src/app/docs/email-validation/email-validation-docs.component.ts
@@ -1,9 +1,0 @@
-import {
-  Component
-} from '@angular/core';
-
-@Component({
-  selector: 'app-email-validation-docs',
-  templateUrl: './email-validation-docs.component.html'
-})
-export class EmailValidationDocsComponent { }