+++ libs/components/validation/src/app/docs/url-validation/url-validation-docs.component.ts
@@ -1,9 +1,0 @@
-import {
-  Component
-} from '@angular/core';
-
-@Component({
-  selector: 'app-url-validation-docs',
-  templateUrl: './url-validation-docs.component.html'
-})
-export class UrlValidationDocsComponent { }