+++ libs/components/validation/src/app/public/public_api.ts
@@ -1,4 +1,0 @@
-export * from './modules/email-validation/email-validation.module';
-export * from './modules/url-validation/url-validation.module';
-export * from './modules/validation/validation';
-export * from './modules/validators/validators';